ytshortsgen_mrbeast/requirements.txt

yt-dlp
openai-whisper
moviepy
openai
python-telegram-bot==13.14
